fix(design-workspace): rollback multi-framework casting pipeline, restore working sandbox renderer
Reverts the project-native architecture (8d8bb7be) that broke the design
workspace preview. Restores the original simple sandbox renderer that
correctly compiles and renders components in an iframe. Recovers the
compile-preview API route that was missing from the rollback.

Removes: framework-renderers, component-casting, worktree-manager,
project-detection, inspector-script, inspector-proxy (~6300 lines).
