+++ lib/ai/tools/design-workspace-tool.ts
@@ -53,83 +53,6 @@ import {
   buildInspectPromptText,
   type InspectMessageContext,
 } from "../../design/workspace/inspect-context";
-import { detectFramework, buildProjectStructure, validateProjectPath } from "@/lib/design/workspace/project-detection";
-import type { FrameworkType } from "@/lib/design/workspace/project-detection";
-import { createDirectDesignWorktree, getActiveWorktree, cleanupDesignWorktree, type WorktreeInfo } from "@/lib/design/workspace/worktree-manager";
-import { castProjectFile, recastAfterEdit, applyEditToWorktree, readWorktreeFile, syncBackChanges } from "@/lib/design/workspace/component-casting";
-import { initializeRenderers } from "@/lib/design/workspace/framework-renderers/init";
-import { rendererRegistry } from "@/lib/design/workspace/framework-renderers/registry";
-import { isPathAllowed, resolveWorkspaceAwarePaths } from "@/lib/ai/filesystem";
-
-// ---------------------------------------------------------------------------
-// Project path authorization
-// ---------------------------------------------------------------------------
-
-/**
- * Gate a model-supplied `projectSourcePath` against synced folder roots.
- * Returns the resolved path if allowed, or an error string if blocked.
- */
-async function authorizeProjectPath(
-  projectPath: string,
-  characterId: string,
-  sessionId: string,
-): Promise<{ ok: true; resolvedPath: string } | { ok: false; error: string }> {
-  try {
-    const allowedPaths = await resolveWorkspaceAwarePaths(characterId, sessionId);
-    if (allowedPaths.length === 0) {
-      return { ok: false, error: "No synced folders configured. Cannot access project paths." };
-    }
-    const resolved = await isPathAllowed(projectPath, allowedPaths);
-    if (!resolved) {
-      return {
-        ok: false,
-        error: `Project path "${projectPath}" is not within any synced folder. Allowed: ${allowedPaths.join(", ")}`,
-      };
-    }
-    return { ok: true, resolvedPath: resolved };
-  } catch (err) {
-    return { ok: false, error: `Path authorization failed: ${err instanceof Error ? err.message : String(err)}` };
-  }
-}
-
-// ---------------------------------------------------------------------------
-// Cast Registry — session-scoped in-memory tracking for cast components
-// Cast components are NOT persisted to the gallery DB; they live only as long
-// as the session is active. This registry makes them discoverable via status,
-// list, and readSource actions.
-// ---------------------------------------------------------------------------
-
-interface CastRegistryEntry {
-  componentId: string;
-  targetFile: string;
-  castMode: "page" | "component" | "route";
-  sourceCode: string;
-  previewHtml: string;
-  framework: string;
-  lastCastAt: number;
-}
-
-const castRegistry = new Map<string, Map<string, CastRegistryEntry>>();
-
-function getCastEntry(sessionId: string, componentId: string): CastRegistryEntry | null {
-  return castRegistry.get(sessionId)?.get(componentId) ?? null;
-}
-
-function setCastEntry(sessionId: string, entry: CastRegistryEntry): void {
-  if (!castRegistry.has(sessionId)) {
-    castRegistry.set(sessionId, new Map());
-  }
-  castRegistry.get(sessionId)!.set(entry.componentId, entry);
-}
-
-function listCastEntries(sessionId: string): CastRegistryEntry[] {
-  const entries = castRegistry.get(sessionId);
-  return entries ? Array.from(entries.values()) : [];
-}
-
-function clearCastRegistry(sessionId: string): void {
-  castRegistry.delete(sessionId);
-}
 
 interface DesignWorkspaceToolOptions {
   sessionId?: string;
@@ -148,7 +71,7 @@ export function resolveComponentCode(_sessionId: string, code: string): string |
 }
 
 interface DesignWorkspaceInput {
-  action: "open" | "generate" | "edit" | "patch" | "readSource" | "list" | "status" | "close" | "install" | "detect" | "browse" | "cast" | "sync-back";
+  action: "open" | "generate" | "edit" | "patch" | "readSource" | "list" | "status" | "close" | "install";
 
   /** npm package names to install. Required for "install" action. */
   packages?: string[];
@@ -178,21 +101,12 @@ interface DesignWorkspaceInput {
    * requires changes at multiple locations in the source.
    */
   patches?: Array<{ oldString: string; newString: string; replaceAll?: boolean }>;
-
-  projectSourcePath?: string;
-  targetFile?: string;
-  castMode?: "page" | "component" | "route";
-  frameworkOverride?: FrameworkType;
-  routePath?: string;
-  syncStrategy?: "merge" | "pr" | "cherry-pick";
-  syncFiles?: string[];
-  browseFilter?: "pages" | "components" | "layouts" | "styles" | "all";
 }
 
 interface ListedDesignSummary {
   id: string;
   name: string;
-  source: "session" | "saved" | "cast";
+  source: "session" | "saved";
   updatedAt?: string;
   isFavorite?: boolean;
 }
@@ -224,17 +138,6 @@ interface DesignWorkspaceResultData {
   };
   recoveryHint?: string;
   updatedAt?: string;
-  metadata?: Record<string, unknown>;
-  /** Project-mode fields */
-  castFile?: string;
-  castMode?: "page" | "component" | "route";
-  rendererInfo?: {
-    type: string;
-    port?: number;
-    baseUrl?: string;
-    /** Full route-qualified preview URL from the renderer. */
-    previewUrl?: string;
-  };
 }
 
 interface DesignWorkspaceResult {
@@ -667,14 +570,11 @@ async function executeDesignWorkspace(
   options: DesignWorkspaceToolOptions,
   input: DesignWorkspaceInput,
 ): Promise<DesignWorkspaceResult> {
-  initializeRenderers();
   let result: DesignWorkspaceResult;
-  const sessionId = getSessionId(options);
-  const config = getWorkspaceConfig();
 
   switch (input.action) {
     case "open":
-      result = await handleOpen(options, input);
+      result = await handleOpen(options);
       break;
     case "install":
       result = await handleInstall(options, input);
@@ -698,19 +598,7 @@ async function executeDesignWorkspace(
       result = await handleStatus(options, input);
       break;
     case "close":
-      result = await handleClose(options);
-      break;
-    case "detect":
-      result = await handleDetect(input, sessionId, options.characterId);
-      break;
-    case "browse":
-      result = await handleBrowse(input, sessionId, options.characterId);
-      break;
-    case "cast":
-      result = await handleCast(input, sessionId, config, options.characterId);
-      break;
-    case "sync-back":
-      result = await handleSyncBack(input, sessionId);
+      result = handleClose(options);
       break;
     default:
       result = { success: false, action: String(input.action), error: `Unknown action: ${input.action}` };
@@ -738,11 +626,7 @@ export function createDesignWorkspaceTool(options: DesignWorkspaceToolOptions = 
 - "readSource": Read back the source code of a persisted component. Pass \`activeComponentId\`.
 - "list": List designs available to the current workspace session.
 - "status": Inspect whether a design is persisted and available. Pass \`activeComponentId\`.
-- "close": Close the design workspace panel.
-- "detect": Detect framework and project structure from a local project path. Provide \`projectSourcePath\`.
-- "browse": Browse pages, components, layouts, and styles in a detected project. Provide \`projectSourcePath\` and optional \`browseFilter\`.
-- "cast": Cast a project file into the design workspace for live preview. Provide \`targetFile\` (relative path), optional \`castMode\` ("page"|"component"|"route"), and \`projectSourcePath\`.
-- "sync-back": Sync worktree changes back to the original project. Provide \`projectSourcePath\` and optional \`syncStrategy\` ("merge"|"pr"|"cherry-pick").`,
+- "close": Close the design workspace panel.`,
     inputSchema: jsonSchema<DesignWorkspaceInput>({
       type: "object",
       title: "DesignWorkspaceInput",
@@ -750,7 +634,7 @@ export function createDesignWorkspaceTool(options: DesignWorkspaceToolOptions = 
       properties: {
         action: {
           type: "string",
-          enum: ["open", "generate", "edit", "patch", "readSource", "list", "status", "close", "install", "detect", "browse", "cast", "sync-back"],
+          enum: ["open", "generate", "edit", "patch", "readSource", "list", "status", "close", "install"],
           description: "The workspace action to perform.",
         },
         packages: {
@@ -827,38 +711,6 @@ export function createDesignWorkspaceTool(options: DesignWorkspaceToolOptions = 
             additionalProperties: false,
           },
         },
-        projectSourcePath: {
-          type: "string",
-          description: "Absolute path to the synced folder / project root. Used with 'detect', 'cast', and 'open' (for project mode).",
-        },
-        targetFile: {
-          type: "string",
-          description: "Target file path relative to project root. Used with 'cast' action.",
-        },
-        castMode: {
-          type: "string",
-          enum: ["page", "component", "route"],
-          description: "How to render the target: as a full page, isolated component, or by route. Used with 'cast'.",
-        },
-        frameworkOverride: {
-          type: "string",
-          description: "Override auto-detected framework type.",
-        },
-        syncStrategy: {
-          type: "string",
-          enum: ["merge", "pr", "cherry-pick"],
-          description: "Strategy for syncing worktree changes back to source. Used with 'sync-back'.",
-        },
-        syncFiles: {
-          type: "array",
-          items: { type: "string" },
-          description: "Specific files to sync back (for cherry-pick strategy).",
-        },
-        browseFilter: {
-          type: "string",
-          enum: ["pages", "components", "layouts", "styles", "all"],
-          description: "Filter for browsing project structure. Used with 'browse'.",
-        },
       },
       required: ["action"],
       additionalProperties: false,
@@ -867,294 +719,7 @@ export function createDesignWorkspaceTool(options: DesignWorkspaceToolOptions = 
   });
 }
 
-// ---------------------------------------------------------------------------
-// Project-native action handlers
-// ---------------------------------------------------------------------------
-
-async function handleDetect(
-  input: DesignWorkspaceInput,
-  sessionId: string,
-  characterId?: string,
-): Promise<DesignWorkspaceResult> {
-  const startedAt = Date.now();
-  ensureHistory(sessionId);
-  const projectPath = input.projectSourcePath;
-  if (!projectPath) {
-    return { success: false, action: "detect", error: "projectSourcePath is required for detect action." };
-  }
-
-  // Authorization: verify projectSourcePath is within synced folder roots
-  const authResult = await authorizeProjectPath(projectPath, characterId ?? "", sessionId);
-  if (!authResult.ok) {
-    return { success: false, action: "detect", error: authResult.error };
-  }
-
-  try {
-    const validation = await validateProjectPath(authResult.resolvedPath);
-    if (!validation.valid) {
-      return { success: false, action: "detect", error: `Invalid project path: ${validation.reason}` };
-    }
-
-    const safePath = authResult.resolvedPath;
-    const framework = await detectFramework(safePath);
-    const structure = await buildProjectStructure(safePath, framework.type);
-
-    recordDesignHistory(sessionId, {
-      action: "open", // recorded as open since detect initializes context
-      durationMs: Date.now() - startedAt,
-      success: true,
-      metadata: { subAction: "detect", projectPath: safePath, framework: framework.type },
-    });
-
-    return {
-      success: true,
-      action: "detect",
-      data: {
-        message: `Detected ${framework.type} project (${framework.buildTool} build, ${framework.cssFramework} CSS, ${framework.packageManager} package manager).`,
-        config: undefined as any, // not needed for detect
-        // Encode project info in metadata-style fields
-        prompt: JSON.stringify({ framework, structure }, null, 2),
-      },
-    };
-  } catch (error) {
-    return {
-      success: false,
-      action: "detect",
-      error: `Detection failed: ${error instanceof Error ? error.message : String(error)}`,
-    };
-  }
-}
-
-async function handleBrowse(
-  input: DesignWorkspaceInput,
-  sessionId: string,
-  characterId?: string,
-): Promise<DesignWorkspaceResult> {
-  ensureHistory(sessionId);
-  const projectPath = input.projectSourcePath;
-  if (!projectPath) {
-    return { success: false, action: "browse", error: "projectSourcePath is required for browse action." };
-  }
-
-  // Authorization: verify projectSourcePath is within synced folder roots
-  const authResult = await authorizeProjectPath(projectPath, characterId ?? "", sessionId);
-  if (!authResult.ok) {
-    return { success: false, action: "browse", error: authResult.error };
-  }
-
-  try {
-    const safePath = authResult.resolvedPath;
-    const framework = await detectFramework(safePath);
-    const structure = await buildProjectStructure(safePath, framework.type);
-    const filter = input.browseFilter ?? "all";
-
-    let entries: Array<{ path: string; name: string; route?: string; type: string }> = [];
-    if (filter === "all" || filter === "pages") entries.push(...structure.pages);
-    if (filter === "all" || filter === "components") entries.push(...structure.components);
-    if (filter === "all" || filter === "layouts") entries.push(...structure.layouts);
-    if (filter === "all" || filter === "styles") entries.push(...structure.styles);
-
-    return {
-      success: true,
-      action: "browse",
-      data: {
-        message: `Found ${entries.length} entries (${structure.pages.length} pages, ${structure.components.length} components, ${structure.layouts.length} layouts, ${structure.styles.length} styles).`,
-        prompt: JSON.stringify(entries, null, 2),
-      },
-    };
-  } catch (error) {
-    return {
-      success: false,
-      action: "browse",
-      error: `Browse failed: ${error instanceof Error ? error.message : String(error)}`,
-    };
-  }
-}
-
-async function handleCast(
-  input: DesignWorkspaceInput,
-  sessionId: string,
-  config: DesignWorkspaceConfig,
-  characterId?: string,
-): Promise<DesignWorkspaceResult> {
-  const startedAt = Date.now();
-  ensureHistory(sessionId);
-  const targetFile = input.targetFile;
-  const castMode = input.castMode ?? "page";
-
-  if (!targetFile) {
-    return { success: false, action: "cast", error: "targetFile is required for cast action." };
-  }
-
-  try {
-    // Ensure worktree exists — if not, create a direct-mode worktree that
-    // points at the real project source.  This avoids duplicating the
-    // directory and lets the dev server use the existing node_modules.
-    let worktree = getActiveWorktree(sessionId);
-    if (!worktree && input.projectSourcePath) {
-      const authResult = await authorizeProjectPath(input.projectSourcePath, characterId ?? "", sessionId);
-      if (!authResult.ok) {
-        return { success: false, action: "cast", error: authResult.error };
-      }
-      worktree = await createDirectDesignWorktree(authResult.resolvedPath, sessionId);
-    }
-    if (!worktree) {
-      return { success: false, action: "cast", error: "No active worktree. Provide projectSourcePath or run 'detect' first." };
-    }
-
-    // Detect framework — pass the full DetectedFramework object so renderers
-    // get buildTool, cssFramework, packageManager etc. instead of just the type.
-    const detected = await detectFramework(worktree.worktreePath);
-    const framework = input.frameworkOverride
-      ? { ...detected, type: input.frameworkOverride }
-      : detected;
-
-    const result = await castProjectFile(sessionId, targetFile, castMode, framework, config);
-
-    // Extract renderer info for C14 — populate RendererInfo in the result.
-    // For dev-server renderers, use the inspector proxy URL so the preview
-    // iframe loads through the proxy (which injects the inspector script).
-    // Store both baseUrl (proxy root) and previewUrl (full route-qualified URL
-    // from the renderer's render() output, which includes framework-specific
-    // route mapping like fileToRoute/bladeToRoute).
-    let rendererInfo: DesignWorkspaceResultData["rendererInfo"];
-    const activeRenderer = rendererRegistry.getActiveRenderer(worktree.worktreePath);
-    if (activeRenderer) {
-      rendererInfo = {
-        type: activeRenderer.tier === "dev-server" ? "dev-server" : "esbuild",
-      };
-      const proxyUrl = (activeRenderer as { inspectorProxyUrl?: string | null }).inspectorProxyUrl;
-      if (proxyUrl) {
-        const port = new URL(proxyUrl).port;
-        rendererInfo.port = parseInt(port, 10);
-        rendererInfo.baseUrl = proxyUrl;
-      }
-      // Carry the full route-qualified preview URL from render() output.
-      // The renderer already applied framework-specific route resolution
-      // (fileToRoute, bladeToRoute, etc.) — the frontend should use this
-      // instead of reconstructing baseUrl + castFile.
-      if (result.rendererProxyUrl) {
-        rendererInfo.previewUrl = result.rendererProxyUrl;
-      }
-    }
-
-    // Register in cast registry so status/list/readSource can find it
-    setCastEntry(sessionId, {
-      componentId: result.componentId,
-      targetFile,
-      castMode,
-      sourceCode: result.sourceCode,
-      previewHtml: result.previewHtml,
-      framework: framework.type,
-      lastCastAt: Date.now(),
-    });
-
-    recordDesignHistory(sessionId, {
-      action: "generate", // cast is conceptually like generate
-      componentId: result.componentId,
-      durationMs: Date.now() - startedAt,
-      success: true,
-      metadata: { subAction: "cast", targetFile, castMode, framework: framework.type },
-    });
-
-    return {
-      success: true,
-      action: "cast",
-      data: {
-        componentId: result.componentId,
-        code: result.sourceCode,
-        previewHtml: result.previewHtml,
-        compileReport: result.compileReport,
-        message: `Cast ${targetFile} as ${castMode} (${framework.type}).`,
-        castFile: targetFile,
-        castMode,
-        rendererInfo,
-      },
-    };
-  } catch (error) {
-    recordDesignHistory(sessionId, {
-      action: "generate",
-      durationMs: Date.now() - startedAt,
-      success: false,
-      error: error instanceof Error ? error.message : String(error),
-      metadata: { subAction: "cast", targetFile, castMode },
-    });
-    return {
-      success: false,
-      action: "cast",
-      error: `Cast failed: ${error instanceof Error ? error.message : String(error)}`,
-    };
-  }
-}
-
-async function handleSyncBack(
-  input: DesignWorkspaceInput,
-  sessionId: string,
-): Promise<DesignWorkspaceResult> {
-  const startedAt = Date.now();
-  ensureHistory(sessionId);
-  const strategy = input.syncStrategy ?? "merge";
-
-  try {
-    const worktree = getActiveWorktree(sessionId);
-    if (!worktree) {
-      return { success: false, action: "sync-back", error: "No active worktree for this session." };
-    }
-
-    // Determine source root — the original project, not the worktree
-    // We need the projectSourcePath or derive it from worktree metadata
-    const sourceRoot = input.projectSourcePath;
-    if (!sourceRoot) {
-      return { success: false, action: "sync-back", error: "projectSourcePath is required for sync-back." };
-    }
-
-    // In direct mode changes are already in the source — sync-back is a no-op.
-    if (worktree.isDirect) {
-      recordDesignHistory(sessionId, {
-        action: "close",
-        durationMs: Date.now() - startedAt,
-        success: true,
-        metadata: { subAction: "sync-back", strategy, directMode: true },
-      });
-      return {
-        success: true,
-        action: "sync-back",
-        data: {
-          message: "Direct mode — all changes are already in the source project. No sync needed.",
-        },
-      };
-    }
-
-    const result = await syncBackChanges(sessionId, sourceRoot, strategy, input.syncFiles);
-
-    recordDesignHistory(sessionId, {
-      action: "close", // sync-back is conceptually a finalization
-      durationMs: Date.now() - startedAt,
-      success: result.success,
-      metadata: { subAction: "sync-back", strategy, appliedFiles: result.appliedFiles },
-    });
-
-    return {
-      success: result.success,
-      action: "sync-back",
-      data: {
-        message: result.success
-          ? `Synced ${result.appliedFiles.length} files back to source (${strategy}).`
-          : `Sync-back failed: ${result.error}`,
-        prompt: result.diff, // Include diff in response
-      },
-      error: result.error,
-    };
-  } catch (error) {
-    return {
-      success: false,
-      action: "sync-back",
-      error: `Sync-back failed: ${error instanceof Error ? error.message : String(error)}`,
-    };
-  }
-}
-
-async function handleOpen(options: DesignWorkspaceToolOptions, input?: DesignWorkspaceInput): Promise<DesignWorkspaceResult> {
+async function handleOpen(options: DesignWorkspaceToolOptions): Promise<DesignWorkspaceResult> {
   const startedAt = Date.now();
   const sessionId = getSessionId(options);
   ensureHistory(sessionId);
@@ -1164,47 +729,9 @@ async function handleOpen(options: DesignWorkspaceToolOptions, input?: DesignWor
   const config = getWorkspaceConfig();
   const history = peekDesignHistory(sessionId);
 
-  // Project-native mode: detect framework and create worktree
-  let metadata: Record<string, unknown> | undefined;
-  if (input?.projectSourcePath) {
-    try {
-      // Authorization: verify projectSourcePath is within synced folder roots
-      const authResult = await authorizeProjectPath(
-        input.projectSourcePath,
-        options.characterId ?? "",
-        sessionId,
-      );
-      if (authResult.ok) {
-        const safePath = authResult.resolvedPath;
-        const validation = await validateProjectPath(safePath);
-        if (validation.valid) {
-          const framework = await detectFramework(safePath);
-          const structure = await buildProjectStructure(safePath, framework.type);
-          // Direct mode: point at the real project source so the dev
-          // server finds existing node_modules.  No directory copy.
-          const worktree = await createDirectDesignWorktree(safePath, sessionId);
-
-          // Include project info in the response
-          metadata = {
-            framework,
-            structure,
-            worktreePath: worktree.worktreePath,
-            worktreeBranch: worktree.branch,
-          };
-        }
-      } else {
-        console.warn("[design-workspace] Project path not authorized:", authResult.error);
-      }
-    } catch (projectError) {
-      // Non-fatal: log but continue with sandbox mode
-      console.warn("[design-workspace] Project detection failed, falling back to sandbox:", projectError);
-    }
-  }
-
   recordHistory(sessionId, "open", startedAt, true, {
     metadata: {
       availableLibraries: available,
-      ...metadata,
     },
   });
 
@@ -1212,11 +739,10 @@ async function handleOpen(options: DesignWorkspaceToolOptions, input?: DesignWor
     success: true,
     action: "open",
     data: {
-      message: metadata ? "Design workspace opened in project mode." : "Design workspace opened.",
+      message: "Design workspace opened.",
       availableLibraries: available.length > 0 ? available : undefined,
       config,
       history: history ?? undefined,
-      metadata,
     },
   };
 }
@@ -1423,12 +949,6 @@ async function handleEdit(
     assets: inputAssets,
   } = input;
 
-  // ── Project-mode branch: edit worktree file instead of gallery design ──
-  const worktree = getActiveWorktree(sessionId);
-  if (worktree && input.targetFile) {
-    return handleEditProjectMode(options, input, sessionId, startedAt, worktree);
-  }
-
   if (!activeComponentId) {
     const error = 'Provide "activeComponentId" to edit a persisted design.';
     recordHistory(sessionId, "edit", startedAt, false, { error });
@@ -1630,12 +1150,6 @@ async function handlePatch(
   ensureHistory(sessionId);
 
   const { oldString, newString, replaceAll: replaceAllOccurrences, activeComponentId, activeComponentCode, patches } = input;
-
-  // ── Project-mode branch: patch worktree file instead of gallery design ──
-  const worktree = getActiveWorktree(sessionId);
-  if (worktree && input.targetFile) {
-    return handlePatchProjectMode(options, input, sessionId, startedAt, worktree);
-  }
 
   if (!activeComponentId) {
     const error = 'Provide "activeComponentId" to patch a persisted design.';
@@ -1856,29 +1370,6 @@ async function handleReadSource(
     };
   }
 
-  // Check cast registry first (cast components aren't in the gallery DB)
-  const castEntry = getCastEntry(sessionId, activeComponentId);
-  if (castEntry) {
-    return {
-      success: true,
-      action: "readSource",
-      data: {
-        componentId: castEntry.componentId,
-        code: castEntry.sourceCode,
-        name: castEntry.targetFile,
-        status: "available",
-        storage: {
-          database: false,
-          userScoped: Boolean(getPersistedUserId(options)),
-          sessionScoped: true,
-        },
-        castFile: castEntry.targetFile,
-        castMode: castEntry.castMode,
-        message: `Cast component source retrieved (${castEntry.sourceCode.length} chars, ~${Math.ceil(castEntry.sourceCode.split("\n").length)} lines).`,
-      },
-    };
-  }
-
   const component = await findWorkspaceDesign({
     id: activeComponentId,
     userId: getPersistedUserId(options),
@@ -1929,41 +1420,24 @@ async function handleList(options: DesignWorkspaceToolOptions): Promise<DesignWo
       limit: 100,
     });
 
-    // Merge in cast registry entries (not in the gallery DB)
-    const castEntries = listCastEntries(sessionId);
-    const castSummaries: ListedDesignSummary[] = castEntries.map((entry) => ({
-      id: entry.componentId,
-      name: entry.targetFile,
-      source: "cast" as const,
-      updatedAt: new Date(entry.lastCastAt).toISOString(),
-    }));
-
-    // Deduplicate: cast entries override DB entries with the same ID
-    const dbIds = new Set(castSummaries.map((c) => c.id));
-    const dbSummaries: ListedDesignSummary[] = components
-      .filter((component) => !dbIds.has(component.id))
-      .map((component) => ({
-        id: component.id,
-        name: component.name,
-        source: (component.sessionId === sessionId ? "session" : "saved") as "session" | "saved",
-        updatedAt: component.updatedAt,
-        isFavorite: component.isFavorite,
-      }));
-
-    const allComponents = [...castSummaries, ...dbSummaries];
-
     recordHistory(sessionId, "list", startedAt, true, {
-      metadata: { count: allComponents.length, castCount: castSummaries.length },
+      metadata: { count: components.length },
     });
 
     return {
       success: true,
       action: "list",
       data: {
-        components: allComponents,
-        message: allComponents.length > 0
-          ? `Found ${allComponents.length} design${allComponents.length === 1 ? "" : "s"} for this workspace (${castSummaries.length} cast, ${dbSummaries.length} persisted).`
-          : "No designs found for this workspace.",
+        components: components.map((component) => ({
+          id: component.id,
+          name: component.name,
+          source: component.sessionId === sessionId ? "session" : "saved",
+          updatedAt: component.updatedAt,
+          isFavorite: component.isFavorite,
+        })),
+        message: components.length > 0
+          ? `Found ${components.length} persisted design${components.length === 1 ? "" : "s"} for this workspace.`
+          : "No persisted designs found for this workspace.",
       },
     };
   } catch (error) {
@@ -2010,31 +1484,6 @@ async function handleStatus(
       success: false,
       action: "status",
       error,
-    };
-  }
-
-  // Check cast registry first (cast components aren't in the gallery DB)
-  const castEntry = getCastEntry(sessionId, input.activeComponentId);
-  if (castEntry) {
-    recordHistory(sessionId, "status", startedAt, true, {
-      componentId: castEntry.componentId,
-    });
-    return {
-      success: true,
-      action: "status",
-      data: {
-        componentId: castEntry.componentId,
-        name: castEntry.targetFile,
-        status: "available",
-        storage: {
-          database: false,
-          userScoped: Boolean(getPersistedUserId(options)),
-          sessionScoped: true,
-        },
-        castFile: castEntry.targetFile,
-        castMode: castEntry.castMode,
-        message: `Cast component from ${castEntry.targetFile} (${castEntry.castMode}).`,
-      },
     };
   }
 
@@ -2089,246 +1538,11 @@ async function handleStatus(
   };
 }
 
-// ---------------------------------------------------------------------------
-// Project-mode edit/patch handlers
-// ---------------------------------------------------------------------------
-
-async function handleEditProjectMode(
-  options: DesignWorkspaceToolOptions,
-  input: DesignWorkspaceInput,
-  sessionId: string,
-  startedAt: number,
-  worktree: WorktreeInfo,
-): Promise<DesignWorkspaceResult> {
-  const { editPrompt, style = "default", activeComponentCode, targetFile, assets: inputAssets } = input;
-
-  if (!targetFile) {
-    return { success: false, action: "edit", error: "targetFile is required for project-mode edit." };
-  }
-
-  // Read current source from worktree
-  let currentCode: string;
-  try {
-    currentCode = await readWorktreeFile(sessionId, targetFile);
-  } catch (err) {
-    const error = `Failed to read worktree file: ${err instanceof Error ? err.message : String(err)}`;
-    recordHistory(sessionId, "edit", startedAt, false, { error });
-    return { success: false, action: "edit", error };
-  }
-
-  // Determine final code: direct replacement or AI-driven edit
-  let finalCode = activeComponentCode?.trim() || "";
-  if (!finalCode) {
-    if (!editPrompt?.trim()) {
-      const error = 'Provide "editPrompt" for AI-driven editing, or provide "activeComponentCode" to directly replace the source.';
-      recordHistory(sessionId, "edit", startedAt, false, { error });
-      return { success: false, action: "edit", error };
-    }
-
-    const assets = inputAssets?.length ? await resolveAssets(inputAssets) : undefined;
-    let editError: string | undefined;
-
-    let enrichedEditPrompt = editPrompt;
-    if (options.inspectContext) {
-      const inspectPromptText = buildInspectPromptText(options.inspectContext);
-      if (inspectPromptText) {
-        enrichedEditPrompt = `${inspectPromptText}\n\n${enrichedEditPrompt}`;
-      }
-    }
-
-    for await (const event of editCard({ code: currentCode, editPrompt: enrichedEditPrompt, style, assets })) {
-      if (event.type === "complete") finalCode = event.content;
-      if (event.type === "error") editError = event.error.message;
-    }
-
-    if (editError || !finalCode.trim()) {
-      const error = editError ?? "Edit produced empty output.";
-      recordHistory(sessionId, "edit", startedAt, false, { error });
-      return { success: false, action: "edit", error };
-    }
-  }
-
-  // Write edited code to worktree and re-cast for preview
-  try {
-    const config = getWorkspaceConfig();
-    const detected = await detectFramework(worktree.worktreePath);
-    const framework = input.frameworkOverride
-      ? { ...detected, type: input.frameworkOverride }
-      : detected;
-    const castResult = await recastAfterEdit(sessionId, targetFile, finalCode.trim(), framework, config);
-
-    recordHistory(sessionId, "edit", startedAt, true, {
-      componentId: castResult.componentId,
-      metadata: { subAction: "project-edit", targetFile },
-    });
-
-    return {
-      success: true,
-      action: "edit",
-      data: {
-        componentId: castResult.componentId,
-        code: finalCode.trim(),
-        name: targetFile,
-        message: `Project file "${targetFile}" edited and re-cast successfully.`,
-        previewHtml: castResult.previewHtml,
-        compileReport: castResult.compileReport,
-      },
-    };
-  } catch (err) {
-    const error = `Project edit failed: ${err instanceof Error ? err.message : String(err)}`;
-    recordHistory(sessionId, "edit", startedAt, false, { error, metadata: { targetFile } });
-    return { success: false, action: "edit", error, data: { code: finalCode.trim() } };
-  }
-}
-
-async function handlePatchProjectMode(
-  _options: DesignWorkspaceToolOptions,
-  input: DesignWorkspaceInput,
-  sessionId: string,
-  startedAt: number,
-  worktree: WorktreeInfo,
-): Promise<DesignWorkspaceResult> {
-  const { oldString, newString, replaceAll: replaceAllOccurrences, targetFile, patches } = input;
-
-  if (!targetFile) {
-    return { success: false, action: "patch", error: "targetFile is required for project-mode patch." };
-  }
-
-  // Build patch operations
-  type PatchOp = { oldString: string; newString: string; replaceAll?: boolean };
-  let patchOps: PatchOp[];
-
-  if (patches && Array.isArray(patches) && patches.length > 0) {
-    for (let i = 0; i < patches.length; i++) {
-      const p = patches[i];
-      if (!p.oldString && p.oldString !== "") {
-        return { success: false, action: "patch", error: `patches[${i}]: "oldString" is required.` };
-      }
-      if (p.newString === undefined || p.newString === null) {
-        return { success: false, action: "patch", error: `patches[${i}]: "newString" is required.` };
-      }
-      if (p.oldString === p.newString) {
-        return { success: false, action: "patch", error: `patches[${i}]: "oldString" and "newString" are identical.` };
-      }
-    }
-    patchOps = patches;
-  } else {
-    if (oldString === undefined || oldString === null) {
-      return { success: false, action: "patch", error: '"oldString" is required for patch action (or provide "patches" array).' };
-    }
-    if (newString === undefined || newString === null) {
-      return { success: false, action: "patch", error: '"newString" is required for patch action.' };
-    }
-    if (oldString === newString) {
-      return { success: false, action: "patch", error: '"oldString" and "newString" are identical.' };
-    }
-    patchOps = [{ oldString, newString, replaceAll: replaceAllOccurrences }];
-  }
-
-  // Read current source from worktree
-  let currentCode: string;
-  try {
-    currentCode = await readWorktreeFile(sessionId, targetFile);
-  } catch (err) {
-    const error = `Failed to read worktree file: ${err instanceof Error ? err.message : String(err)}`;
-    recordHistory(sessionId, "patch", startedAt, false, { error });
-    return { success: false, action: "patch", error };
-  }
-
-  // Apply patches sequentially
-  let patchedCode = currentCode;
-  let totalReplacements = 0;
-
-  for (let i = 0; i < patchOps.length; i++) {
-    const op = patchOps[i];
-    const occurrences = patchedCode.split(op.oldString).length - 1;
-
-    if (occurrences === 0) {
-      const patchLabel = patchOps.length > 1 ? ` (patches[${i}])` : "";
-      return {
-        success: false,
-        action: "patch",
-        error: `"oldString" not found in project file${patchLabel}.${i > 0 ? ` Note: ${i} prior patch(es) were already applied.` : ""}`,
-      };
-    }
-
-    if (occurrences > 1 && !op.replaceAll) {
-      const patchLabel = patchOps.length > 1 ? ` (patches[${i}])` : "";
-      return {
-        success: false,
-        action: "patch",
-        error: `"oldString" found ${occurrences} times${patchLabel}. Set "replaceAll: true" to replace all.`,
-      };
-    }
-
-    patchedCode = op.replaceAll
-      ? patchedCode.split(op.oldString).join(op.newString)
-      : patchedCode.replace(op.oldString, op.newString);
-    totalReplacements += op.replaceAll ? occurrences : 1;
-  }
-
-  // JSX balance check
-  const unclosedTag = findUnclosedJsxTag(patchedCode);
-  if (unclosedTag) {
-    return {
-      success: false,
-      action: "patch",
-      error: `Patch produced unbalanced JSX: <${unclosedTag}> appears to be unclosed.`,
-    };
-  }
-
-  // Write patched code to worktree and re-cast
-  try {
-    const config = getWorkspaceConfig();
-    const detected = await detectFramework(worktree.worktreePath);
-    const framework = input.frameworkOverride
-      ? { ...detected, type: input.frameworkOverride }
-      : detected;
-    const castResult = await recastAfterEdit(sessionId, targetFile, patchedCode, framework, config);
-
-    const linesChanged = countChangedLines(currentCode, patchedCode);
-    recordHistory(sessionId, "patch", startedAt, true, {
-      componentId: castResult.componentId,
-      metadata: { subAction: "project-patch", targetFile, totalReplacements },
-    });
-
-    return {
-      success: true,
-      action: "patch",
-      data: {
-        componentId: castResult.componentId,
-        code: patchedCode,
-        name: targetFile,
-        message: `Project file "${targetFile}" patched: ${totalReplacements} replacement${totalReplacements > 1 ? "s" : ""}, ~${linesChanged} line${linesChanged !== 1 ? "s" : ""} changed.`,
-        previewHtml: castResult.previewHtml,
-        compileReport: castResult.compileReport,
-      },
-    };
-  } catch (err) {
-    const error = `Project patch failed: ${err instanceof Error ? err.message : String(err)}`;
-    recordHistory(sessionId, "patch", startedAt, false, { error, metadata: { targetFile } });
-    return { success: false, action: "patch", error, data: { code: patchedCode } };
-  }
-}
-
-async function handleClose(options: DesignWorkspaceToolOptions): Promise<DesignWorkspaceResult> {
+function handleClose(options: DesignWorkspaceToolOptions): DesignWorkspaceResult {
   const startedAt = Date.now();
   const sessionId = getSessionId(options);
   ensureHistory(sessionId);
 
-  // Shut down any active renderer for this session's worktree, then clean up the worktree
-  try {
-    const worktree = getActiveWorktree(sessionId);
-    if (worktree) {
-      // Kill dev-server / renderer processes before removing the worktree directory
-      try { await rendererRegistry.shutdownRenderer(worktree.worktreePath); } catch { /* non-fatal */ }
-      await cleanupDesignWorktree(sessionId);
-    }
-  } catch {
-    // Non-fatal
-  }
-
-  clearCastRegistry(sessionId);
   recordHistory(sessionId, "close", startedAt, true);
   const history = finalizeDesignHistory(sessionId);
