+++ lib/design/workspace/dependencies.ts
@@ -267,139 +267,3 @@ export async function installSandboxPackages(
     packageNames,
   };
 }
-
-/**
- * Install packages in a project worktree directory.
- *
- * Unlike `installSandboxPackages` which installs into the design sandbox,
- * this installs into the project's own node_modules within the worktree.
- * Respects the project's package manager (npm, yarn, pnpm, bun).
- */
-export async function installProjectPackages(
-  worktreePath: string,
-  packages: string[],
-): Promise<DependencyInstallResult> {
-  const specs = uniqueStrings(
-    packages
-      .map((packageName) => validatePackageSpec(packageName))
-      .filter((result) => result.valid)
-      .map((result) => result.spec),
-  );
-
-  if (specs.length === 0) {
-    return {
-      attempted: false,
-      success: true,
-      packages: [],
-      packageNames: [],
-    };
-  }
-
-  const packageNames = extractPackageNames(specs);
-
-  try {
-    // Detect package manager from lockfiles
-    const pm = detectPackageManager(worktreePath);
-    const installCmd = getInstallCommand(pm, specs);
-
-    const { command, options: platformOptions } = pm === "npm"
-      ? getNpmExecConfig()
-      : { command: pm, options: IS_WINDOWS ? { shell: true } as ExecFileOptions : {} };
-
-    await execFileAsync(
-      command,
-      installCmd.args,
-      {
-        cwd: worktreePath,
-        timeout: 180_000,
-        env: { ...process.env, NODE_ENV: "development" },
-        ...platformOptions,
-      },
-    );
-
-    return {
-      attempted: true,
-      success: true,
-      packages: specs,
-      packageNames,
-    };
-  } catch (error) {
-    return {
-      attempted: true,
-      success: false,
-      packages: specs,
-      packageNames,
-      error: error instanceof Error ? error.message : String(error),
-    };
-  }
-}
-
-/** Detect which package manager a project uses based on lockfiles. */
-function detectPackageManager(projectRoot: string): "npm" | "yarn" | "pnpm" | "bun" {
-  if (existsSync(join(projectRoot, "pnpm-lock.yaml"))) return "pnpm";
-  if (existsSync(join(projectRoot, "yarn.lock"))) return "yarn";
-  if (existsSync(join(projectRoot, "bun.lockb")) || existsSync(join(projectRoot, "bun.lock"))) return "bun";
-  return "npm";
-}
-
-/** Build the install command args for a given package manager. */
-function getInstallCommand(pm: "npm" | "yarn" | "pnpm" | "bun", specs: string[]): { args: string[] } {
-  switch (pm) {
-    case "pnpm":
-      return { args: ["add", ...specs] };
-    case "yarn":
-      return { args: ["add", ...specs] };
-    case "bun":
-      return { args: ["add", ...specs] };
-    case "npm":
-    default:
-      return { args: ["install", "--save", "--ignore-scripts", ...specs] };
-  }
-}
-
-/**
- * Validate that imported packages resolve in the project's node_modules.
- * Similar to validateWorkspaceDependencies but checks project context.
- */
-export async function validateProjectDependencies(
-  componentCode: string,
-  projectNodeModulesPath: string,
-): Promise<DependencyValidationResult> {
-  const importedPackages = extractImportedPackages(componentCode);
-
-  // Read project's package.json for manifest
-  let manifestPackages: string[] = [];
-  const projectPkgJsonPath = join(projectNodeModulesPath, "..", "package.json");
-  try {
-    const raw = await fs.readFile(projectPkgJsonPath, "utf8");
-    const parsed = JSON.parse(raw) as { dependencies?: Record<string, string>; devDependencies?: Record<string, string> };
-    manifestPackages = [
-      ...Object.keys(parsed.dependencies ?? {}),
-      ...Object.keys(parsed.devDependencies ?? {}),
-    ];
-  } catch {
-    // No package.json or invalid
-  }
-
-  const missingImportedPackages = importedPackages.filter(
-    (packageName) => !canResolvePackage(packageName, [projectNodeModulesPath]),
-  );
-
-  const missingManifestPackages = manifestPackages.filter(
-    (packageName) => !canResolvePackage(packageName, [projectNodeModulesPath]),
-  );
-
-  const missingPackages = uniqueStrings([
-    ...missingManifestPackages,
-    ...missingImportedPackages,
-  ]);
-
-  return {
-    manifestPackages,
-    importedPackages,
-    checkedPackages: uniqueStrings([...manifestPackages, ...importedPackages]),
-    missingManifestPackages,
-    missingImportedPackages,
-    missingPackages,
-  };
-}
