+++ lib/design/workspace/index.ts
@@ -1,17 +1,2 @@
 export { useDesignWorkspaceStore } from "./store";
 export { DESIGN_BREAKPOINTS } from "./types";
-export type {
-  DetectedFramework,
-  FrameworkType,
-  ProjectEntry,
-  ProjectStructure,
-  ProjectContext,
-  DesignComponent,
-  DesignSnapshot,
-  DesignBreakpoint,
-  DesignWorkspaceStatus,
-  InspectedElement,
-  RendererInfo,
-  DesignWorkspaceSessionState,
-  DesignWorkspaceState,
-} from "./types";
